Double speed of BSF
Commented out the assert line in the BSF and BSR functions to speed up the function calls.

+++ majikthise/board.py
@@ -29,7 +29,7 @@ class Piece(IntEnum):
 	K = 5
 
 class Square(IntEnum):
-	A1 = 0 
+	A1 = 0
 	B1 = 1
 	C1 = 2
 	D1 = 3
@@ -39,64 +39,64 @@ class Square(IntEnum):
 	H1 = 7
 	A2 = 8
 	B2 = 9
-	C2 = 10 
-	D2 = 11 
-	E2 = 12 
-	F2 = 13 
+	C2 = 10
+	D2 = 11
+	E2 = 12
+	F2 = 13
 	G2 = 14
 	H2 = 15
-	A3 = 16 
-	B3 = 17 
-	C3 = 18 
-	D3 = 19 
-	E3 = 20 
-	F3 = 21 
-	G3 = 22 
+	A3 = 16
+	B3 = 17
+	C3 = 18
+	D3 = 19
+	E3 = 20
+	F3 = 21
+	G3 = 22
 	H3 = 23
-	A4 = 24 
-	B4 = 25 
-	C4 = 26 
-	D4 = 27 
-	E4 = 28 
-	F4 = 29 
-	G4 = 30 
+	A4 = 24
+	B4 = 25
+	C4 = 26
+	D4 = 27
+	E4 = 28
+	F4 = 29
+	G4 = 30
 	H4 = 31
-	A5 = 32 
-	B5 = 33 
-	C5 = 34 
-	D5 = 35 
-	E5 = 36 
-	F5 = 37 
-	G5 = 38 
+	A5 = 32
+	B5 = 33
+	C5 = 34
+	D5 = 35
+	E5 = 36
+	F5 = 37
+	G5 = 38
 	H5 = 39
-	A6 = 40 
-	B6 = 41 
-	C6 = 42 
-	D6 = 43 
-	E6 = 44 
-	F6 = 45 
-	G6 = 46 
+	A6 = 40
+	B6 = 41
+	C6 = 42
+	D6 = 43
+	E6 = 44
+	F6 = 45
+	G6 = 46
 	H6 = 47
-	A7 = 48 
-	B7 = 49 
-	C7 = 50 
-	D7 = 51 
-	E7 = 52 
-	F7 = 53 
-	G7 = 54 
+	A7 = 48
+	B7 = 49
+	C7 = 50
+	D7 = 51
+	E7 = 52
+	F7 = 53
+	G7 = 54
 	H7 = 55
-	A8 = 56 
-	B8 = 57 
-	C8 = 58 
-	D8 = 59 
-	E8 = 60 
-	F8 = 61 
-	G8 = 62 
+	A8 = 56
+	B8 = 57
+	C8 = 58
+	D8 = 59
+	E8 = 60
+	F8 = 61
+	G8 = 62
 	H8 = 63
 	NONE = 64
 
 	def bitboard(self):
-		return np.uint64(0x01) << np.uint64(self.value)
+		return np.uint64(0x01) << np.uint64(self.value)			
 
 	def isEmpty(self, board: 'CBoard'):
 		# Returns true if no white or black pieces occupy the square.
@@ -138,7 +138,8 @@ class Fen:
 				squareIdx += 1
 			elif char.isnumeric():
 				idx += 1
-				squareIdx += np.uint64(int(char))
+				squareIdx += int(char)
+				#squareIdx += np.uint64(int(char))
 			else:
 				idx += 1
 				squareIdx += 1
